requirements.txt
# Deep Learning Framework
tensorflow
keras

# Data Processing
numpy
pandas

# Image Processing
Pillow
opencv-python

# Machine Learning
scikit-learn

# Visualization
matplotlib
seaborn

# Jupyter
jupyter
notebook
ipywidgets

# Utilities
tqdm